test(reconciliation): split logical revision coverage
